backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
alembic>=1.13.0
httpx>=0.26.0
